Add files via upload
This is the Hellsing-Cipher program. It uses 256AES encryption. It offers you the ability to encrypt and decrypt messages, files and folders/directories. It asks for a password twice, in order to ensure the right password is used. 
Here, we've included the raw python file and the executable linux file, for your convenience.