Add files via upload
This hellsing_cipher v11 program was made to run on windows. It was compiled on Win 11. WARNING: Because it is not officially signed AND contains encryption coding, Windows will flag is as Ransomware, Malware or an encryption virus. I assure you, it's not. Make sure you check the checksum to ensure you downloaded the proper file. Also, if you're still a little sus, download the Python Raw file, go over the coding yourself and you'll see it's A-ok. Oh, that being said, we're not responsible for any damages of any kind that come from your use of this software. It's suggested that you make backups of your files before encrypting, just in case.